full-requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
comtypes==1.4.10
distlib==0.3.9
distro==1.9.0
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.0
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.5.0
networkx==3.5
numpy==2.2.5
openai==0.28.0
packaging==25.0
platformdirs==4.3.8
pocketsphinx==5.0.4
propcache==0.3.2
PyAudio==0.2.14
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyjokes==0.8.3
pypiwin32==223
pyttsx3==2.98
pywin32==310
PyYAML==6.0.2
redis==6.2.0
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
setuptools==80.9.0
sniffio==1.3.1
sounddevice==0.5.2
soupsieve==2.7
SpeechRecognition==3.14.3
spotipy==2.25.1
sympy==1.14.0
tokenizers==0.21.1
torch==2.7.1
tqdm==4.67.1
transformers==4.52.4
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
virtualenv==20.31.2
wikipedia==1.4.0
yarl==1.20.1
langchain
langchain-google-genai
langchain-community
speechrecognition
pyttsx3
spotipy
python-dotenv
pyjokes
wikipedia
requests
transformers
sentence-transformers
chromadb
pyaudio 